lung_dx/requirements.txt
# ── Core ──────────────────────────────────────────────────────
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
openpyxl>=3.1

# ── Settings ──────────────────────────────────────────────────
pydantic>=2.0
pydantic-settings>=2.0

# ── ML / Image ────────────────────────────────────────────────
torch>=2.0
torchvision>=0.15
Pillow>=10.0
grad-cam>=1.5

# ── API ───────────────────────────────────────────────────────
fastapi>=0.100
uvicorn>=0.20

# ── AWS (점진 확장) ───────────────────────────────────────────
boto3>=1.28

# ── ETL ───────────────────────────────────────────────────────
pyarrow>=14.0
